Added C14 prices #15
C14 has a peculiar setup where bandwidth is free, but you pay for operations *in GB* to/from a scratch storage area called your "safe-deposit box".

Since it's not possible to access data without going through the "safe-deposit box", I've put this up as bandwidth costs.  However, since it is possible to remove data from the "safe-deposit box" and not pass it on to storage, it is possible to pay less than the operations cost, therefore the range is 0-1.0 for example, and not 1.0.

The prices are in EUR.  Not sure how to write euro cents, but I've used €c.